interface circle

+++ src/fr/interfaces/Circle.java
@@ -1,8 +1,8 @@
 package fr.interfaces;
 
 public interface Circle {
-	public int getRadius();
-	public int getPositionX();
-	public int getPositionY();
+	public double getRadius();
+	public double getPositionX();
+	public double getPositionY();
 
 }
